requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flasgger
requests
python-dotenv
PyJWT
cryptography
gunicorn
ofxparse
